+++ script/core/diagnostics/helper/missing-doc-helper.lua
@@ -1,6 +1,6 @@
-local lang    = require 'language'
+local lang = require 'language'
 
-local m = {}
+local m    = {}
 
 local function findParam(docs, param)
     if not docs then
@@ -36,14 +36,25 @@ local function findReturn(docs, index)
     return false
 end
 
-local function checkFunction(source, callback, commentId, paramId, returnId)
-    local functionName = source.parent[1]
+---@param functionName string
+---@param source parser.object
+---@param diagnosticRangeSource? parser.object sometimes the object with the data isn't the one that needs the diagnostics
+---@param bindDocsSource? parser.object sometimes the object with the bind docs isn't the value (`a.b = c` syntax)
+---@param callback fun(result: any)
+---@param commentId string
+---@param paramId string
+---@param returnId string
+local function checkFunctionNamed(functionName, source, diagnosticRangeSource, bindDocsSource, callback, commentId,
+                                  paramId, returnId)
+    diagnosticRangeSource = diagnosticRangeSource or source
+    bindDocsSource = bindDocsSource or source
     local argCount = source.args and #source.args or 0
+    local noRealArgs = argCount == 0 or (argCount == 1 and source.args[1][1] == 'self')
 
-    if argCount == 0 and not source.returns and not source.bindDocs then
+    if noRealArgs and not source.returns and not bindDocsSource.bindDocs then
         callback {
-            start   = source.start,
-            finish  = source.finish,
+            start   = diagnosticRangeSource.start,
+            finish  = diagnosticRangeSource.finish,
             message = lang.script(commentId, functionName),
         }
     end
@@ -53,7 +64,7 @@ local function checkFunction(source, callback, commentId, paramId, returnId)
             local argName = arg[1]
             if  argName ~= 'self'
             and argName ~= '_' then
-                if not findParam(source.bindDocs, argName) then
+                if not findParam(bindDocsSource.bindDocs, argName) then
                     callback {
                         start   = arg.start,
                         finish  = arg.finish,
@@ -67,7 +78,7 @@ local function checkFunction(source, callback, commentId, paramId, returnId)
     if source.returns then
         for _, ret in ipairs(source.returns) do
             for index, expr in ipairs(ret) do
-                if not findReturn(source.bindDocs, index) then
+                if not findReturn(bindDocsSource.bindDocs, index) then
                     callback {
                         start   = expr.start,
                         finish  = expr.finish,
@@ -79,5 +90,28 @@ local function checkFunction(source, callback, commentId, paramId, returnId)
     end
 end
 
+---@param source parser.object
+---@param callback fun(result: any)
+---@param commentId string
+---@param paramId string
+---@param returnId string
+local function checkFunction(source, callback, commentId, paramId, returnId)
+    local functionName = source.parent[1]
+    checkFunctionNamed(functionName, source, nil, nil, callback, commentId, paramId, returnId)
+end
+
+
+---@param source parser.object
+---@param callback fun(result: any)
+---@param commentId string
+---@param paramId string
+---@param returnId string
+local function checkMethod(source, callback, commentId, paramId, returnId)
+    local functionName = source.method[1]
+    checkFunctionNamed(functionName, source.value, source, nil, callback, commentId, paramId, returnId)
+end
+
 m.CheckFunction = checkFunction
+m.CheckFunctionNamed = checkFunctionNamed
+m.CheckMethod = checkMethod
 return m